// src/app/api/delivery/seller/requests/mappers/errors/accept-request/accept-request-error-codes.enum.ts
export enum ACCEPT_REQUEST_ERROR_CODES {
  NON_PURCHASABLE_ITEM = 'non purchasable item',
  SELLER_ADDRESS_NOT_FOUND = 'seller address not found',
  ALREADY_IN_PROGRESS_TRANSACTION = 'already in progress transaction',
  SELLER_BLOCK_FOR_FRAUD = 'seller blocked for fraud',
  BUYER_BLOCK_FOR_FRAUD = 'buyer blocked for fraud',
  INVALID_CARD = 'invalid card',
  SELLER_BLOCKED_BY_BUYER = 'seller blocked by buyer',
  BUYER_BLOCKED_BY_SELLER = 'buyer blocked by seller',
  SELLER_ADDRESS_NOT_ALLOWED = 'address not allowed',
  REQUEST_NOT_FOUND = 'request not found',
  POSTAL_CODE_NOT_FOUND = 'postal code not found',
  POSTAL_CODE_NOT_ALLOWED = 'postal code is not allowed',
  POSTAL_CODE_TEMPORARILY_RESTRICTED = 'código postal temporalmente restringido',
  IS_NOT_PENDING_REQUEST = 'not pending request',
  CARRIER_ADDRESSES_DONT_MATCH = `carrier office address and home address countries don't match`,
  NOT_SHIPPABLE_ITEM = 'not shippable item',
  INTERNATIONAL_SHIPPING_NOT_ENABLED = 'international shipping is not enabled for user',
}